+++ src/main/java/com/project/moflis/profile/dto/ProfileResponseDTO.java
@@ -8,8 +8,8 @@ import org.springframework.web.multipart.MultipartFile;
 @Getter
 public class ProfileResponseDTO {
 
-    private Integer id;
-    private Integer userId;
+    private Long id;
+    private Long userId;
     private String intro;
     private MultipartFile profileImage;
     private String profileImageName;
