+++ src/main/java/com/project/moflis/schedules/command/UpdateSchedulsCommand.java
@@ -13,9 +13,9 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class UpdateSchedulsCommand {
 
-    private final Integer scheduleId;
-    private final Integer userId;
-    private final Integer recurringSchedulesId;
+    private final Long scheduleId;
+    private final Long userId;
+    private final Long recurringSchedulesId;
     private final LocalDate scheduleDate;
     private final LocalDateTime startTime;
     private final LocalDateTime endTime;
