+++ src/main/java/com/project/moflis/schedules/controller/SchedulesController.java
@@ -9,14 +9,7 @@ import com.project.moflis.schedules.dto.schedules.UpdateSchedulsRequest;
 import com.project.moflis.schedules.service.SchedulesService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/v1/schedules")
@@ -58,14 +51,14 @@ public class SchedulesController {
 
     //스케줄 수정
     @PatchMapping("/{scheduleId}")
-    public ResponseEntity<ScheduleResponse> patchSchedules(@PathVariable("scheduleId") int scheduleId, UpdateSchedulsRequest request) {
+    public ResponseEntity<ScheduleResponse> patchSchedules(@PathVariable("scheduleId") long scheduleId, UpdateSchedulsRequest request) {
         ScheduleResponse response = schedulsService.updateSchedules(request.toCommand(scheduleId));
         return ResponseEntity.ok(response);
     }
 
     //스케줄 삭제
     @DeleteMapping("/{scheduleId}")
-    public ResponseEntity<Void> deleteSchedules(@PathVariable("scheduleId") int scheduleId) {
+    public ResponseEntity<Void> deleteSchedules(@PathVariable("scheduleId") long scheduleId) {
         schedulsService.deleteSchedules(scheduleId);
         return ResponseEntity.noContent().build();
     }
